Container "finish"? + UserEditGui work
I think (keyword is THINK) the container stuff may work? I need to process the GUIHandler tho as it is not added to the mod yet in the main class.
Need to also finish the GUI for editing the users, since it needs to be capable of writing and stuff yeah