GUI Massive Progress
Got a lot of the GUI stuff working and a TON of gui icons rendering. One thing that will need to be done tho is changing how buttons are rendered, mainly for the user list and stuff. Also a scrollbar might be handy + changing some buttons in user list to have a label and a check/X button instead.